Default and Parameterized Constructor
Constructor is a member function of a class, whose name is the same as the class name. Constructor is a special type of member function that is used to initialize the data members for an object of a class automatically when an object of the same class is created. Constructor is invoked at the time of object creation.